+++ test/extraction.jl
@@ -5,6 +5,8 @@ using PromptingTools: tool_call_signature, set_properties_strict!,
 using PromptingTools: Tool, isabstracttool, execute_tool, parse_tool, get_arg_names,
                       get_arg_types, get_method, get_function, remove_field!,
                       tool_call_signature
+using PromptingTools: AbstractToolError, ToolNotFoundError, ToolExecutionError,
+                      ToolGenericError
 
 # TODO: check more edge cases like empty structs
 
@@ -17,6 +19,22 @@ function context_test_function(x::Int, y::String, ctx_z::Float64)
 end
 function context_test_function2(x::Int, y::String, context::Dict)
     return "Context test: $x, $y, $(context)"
+end
+
+@testset "ToolErrors" begin
+    e = ToolNotFoundError("Tool `xyz` not found")
+    @test e isa AbstractToolError
+    @test e.msg == "Tool `xyz` not found"
+
+    e = ToolExecutionError(
+        "Tool `xyz` execution failed", MethodError(my_test_function, (1,)))
+    @test e isa AbstractToolError
+    @test e.msg == "Tool `xyz` execution failed"
+
+    e = ToolGenericError(
+        "Tool `xyz` failed with a generic error", MethodError(my_test_function, (1,)))
+    @test e isa AbstractToolError
+    @test e.msg == "Tool `xyz` failed with a generic error"
 end
 
 @testset "Tool-constructor" begin
@@ -804,8 +822,12 @@ end
     # Test with missing argument in both args and context
     args_missing = Dict(:x => 5)
     context_missing = Dict(:y => "hello")
-    @test_throws MethodError execute_tool(
+    @test_throws ToolExecutionError execute_tool(
         context_test_function, args_missing, context_missing)
+    err = execute_tool(
+        context_test_function, args_missing, context_missing; throw_on_error = false)
+    @test err isa ToolExecutionError
+    @test err.err isa MethodError
 
     # Test with Tool
     context_tool = Tool(context_test_function)
@@ -818,4 +840,8 @@ end
         tool_call_id = "1", name = "my_test_function", raw = "", args = args)
     output = execute_tool(tool_map, msg)
     @test output == "Test function: 10, hello"
+    ## Call wrong tool name
+    @test_throws ToolNotFoundError execute_tool(tool_map,
+        ToolMessage(;
+            tool_call_id = "1", name = "wrong_tool_name", raw = "", args = args))
 end